# docs/api.yaml
openapi: '3.0.3'
info:
  title: TaskBridge API
  description: REST API маркетплейса digital-услуг TaskBridge (MVP)
  version: '1.0.0'
  contact:
    name: TaskBridge Support
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        category: { type: string, enum: [sites, design, ads, bots, automation, texts, presentations, analytics, other] }
        budget: { type: number }
        budget_type: { type: string, enum: [fixed, negotiable, hourly] }
        deadline: { type: string, format: date }
        urgency: { type: string, enum: [low, normal, high, urgent] }
        status: { type: string, enum: [draft, published, assigned, in_progress, review, completed, cancelled, hidden] }
        creator_id: { type: integer }
        assigned_to: { type: integer, nullable: true }
        required_skills: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
...
paths:
...
  /api/tasks:
    get:
      summary: Список задач
      tags: [Tasks]
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, budget_desc, budget_asc, deadline, least_offers] }
        - name: status
          in: query
          schema: { type: string, default: published }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks: { type: array, items: { $ref: '#/components/schemas/Task' } }
                  total: { type: integer }

    post:
      summary: Создать задачу
      tags: [Tasks]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, category]
              properties:
                title: { type: string }
                description: { type: string }
                category: { type: string }
                budget: { type: number }
                budget_type: { type: string, enum: [fixed, negotiable, hourly] }
                deadline: { type: string }
                urgency: { type: string }
                status: { type: string, enum: [draft, published] }
                required_skills: { type: array, items: { type: string } }
      responses:
        '201': { description: Задача создана }
        '400': { description: Ошибка валидации }
        '401': { description: Не авторизован }
        '403': { description: Только заказчик }
...
  /api/reviews:
    post:
      summary: Оставить отзыв
      tags: [Reviews]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_id, rating]
              properties:
                task_id: { type: integer }
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
      responses:
        '201': { description: Отзыв создан }
        '409': { description: Отзыв уже оставлен }
...
  /api/admin/users:
    get:
      summary: Список пользователей
      tags: [Admin]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Список пользователей }
...